Add resumeProcessing functionality
This means that if a process() fails, a file is created which saves the corresponding Dataset/DataSubset object, along with the remaining subjects and remaining context parameters which had to be processed. Then, the resumeProcessing() method can be use to pick up from where processing failed after the issue has been fixed.

This could be some sort of issue with the raw marker/grf files, like NaN entries, for example.