Core functionality complete
Dataset working
DataSubset working
Resume functionality working

It is now possible to set up a Dataset or DataSubset and run all the OpenSim analyses e.g. IK, RRA, ID, BK.

Some more functionality to be added. For example, currently it is not possible to run all the IK's, go away, then come back and do all the RRA's later from a 'new' Dataset, even though the files are all there. Some saving needs to be put in place - similar to what is there for resume.

Also, the parfor loop is not currently being used to its full extent, but I don't think this is a big problem. Even with only 2 context parameters - 'foot' and 'assistance' say, with foot = 1:2 and assistance = 1:3, a single subject would consume all 6 cores of my work PC. Typically there will be more subjects than this, and an additional context parameter, or more assistance values.